Move specs to correct folder
