+++ specs/client/reducers/decksSpec.js
@@ -1,7 +1,7 @@
 /* global describe, it, before, beforeEach, after, afterEach */
 import { expect } from 'chai';
-import { decks } from '../../../client/reducers';
-import { RECEIVE_DECKS } from '../../../client/constants/actionTypes';
+import { decks } from '../../../src/client/reducers';
+import { RECEIVE_DECKS } from '../../../src/client/constants/actionTypes';
 
 describe('Decks Reducer', () => {
   describe('RECEIVE_DECKS', () => {
