+++ specs/client/reducers/playSpec.js
@@ -1,8 +1,8 @@
 /* global describe, it, before, beforeEach, after, afterEach */
 import { expect } from 'chai';
-import { play } from '../../../client/reducers';
-import { START_PLAY, FLIP_CARD, FINISH_PLAY } from '../../../client/constants/actionTypes';
-import { FRONT, BACK, GREAT, BAD } from '../../../client/constants/play';
+import { play } from '../../../src/client/reducers';
+import { START_PLAY, FLIP_CARD, FINISH_PLAY } from '../../../src/client/constants/actionTypes';
+import { FRONT, BACK, GREAT, BAD } from '../../../src/client/constants/play';
 
 describe('Play Reducer', () => {
   it('should be a function', () => {
